+++ src/tw-styled/style-to-class-resolver/buckets/index.ts
@@ -1,0 +1,6 @@
+export * from "./color";
+export * from "./dimension";
+export * from "./display";
+export * from "./flex";
+export * from "./typography";
+export * from "./spacing";
